Add :core:model, and move to :core:ui
