+++ Compose-Flight-Search-App/core/ui/build.gradle.kts
@@ -19,10 +19,15 @@ android {
 }
 
 dependencies {
+    api(projects.core.model)
 
-    implementation(libs.androidx.core.ktx)
-    implementation(libs.androidx.appcompat)
-    implementation(libs.material)
+    implementation(platform(libs.androidx.compose.bom))
+    implementation(libs.androidx.compose.runtime)
+    implementation(libs.androidx.material3)
+    implementation(libs.androidx.ui.tooling.preview)
+
+    implementation(projects.core.designsystem)
+
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
