+++ Compose-Flight-Search-App/core/ui/src/main/kotlin/com/example/flightsearchapp/core/ui/suggestion/SuggestionsListScreen.kt
@@ -1,4 +1,4 @@
-package com.example.flightsearchapp.ui.suggestion
+package com.example.flightsearchapp.core.ui.suggestion
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -8,8 +8,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.flightsearchapp.ui.model.SuggestionAirport
-import com.example.flightsearchapp.ui.theme.FlightSearchAppTheme
+import com.example.flightsearchapp.core.designsystem.theme.FlightSearchAppTheme
+import com.example.flightsearchapp.core.model.SuggestionAirport
 
 @Composable
 fun SuggestionsListScreen(
