+++ Compose-Flight-Search-App/core/ui/src/main/kotlin/com/example/flightsearchapp/core/ui/uistate/SearchScreenUiState.kt
@@ -1,0 +1,11 @@
+package com.example.flightsearchapp.core.ui.uistate
+
+import com.example.flightsearchapp.core.model.Flight
+import com.example.flightsearchapp.core.model.SuggestionAirport
+
+sealed interface SearchScreenUiState {
+    data class ShowFavorite(val results: List<Flight>) : SearchScreenUiState
+    data class ShowSuggests(val results: List<SuggestionAirport>) :
+        SearchScreenUiState
+    data object Init : SearchScreenUiState
+}
