+++ Compose-Flight-Search-App/core/ui/src/main/kotlin/com/example/flightsearchapp/core/ui/uistate/ShowFlightUiState.kt
@@ -1,6 +1,6 @@
 package com.example.flightsearchapp.core.ui.uistate
 
-import com.example.flightsearchapp.core.ui.model.Flight
+import com.example.flightsearchapp.core.model.Flight
 
 sealed interface ShowFlightUiState {
     data object Clear : ShowFlightUiState
