--- Compose-Flight-Search-App/core/ui/src/main/kotlin/com/example/flightsearchapp/ui/uistate/SearchScreenUiState.kt
@@ -1,7 +1,0 @@
-package com.example.flightsearchapp.ui.uistate
-
-sealed interface SearchScreenUiState {
-    data class ShowFavorite(val results: List<com.example.flightsearchapp.ui.model.Flight>) : SearchScreenUiState
-    data class ShowSuggests(val results: List<com.example.flightsearchapp.data.model.SuggestionAirport>) : SearchScreenUiState
-    data object Init : SearchScreenUiState
-}
